requirements.txt
flask
nltk
numpy
python-dotenv